requirements.txt
aiogram==2.15
psycopg2-binary
python-dotenv
selenium
